requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.11.1
langchain>=0.3.9
langchain-core>=0.3.34
langchain-anthropic>=0.3.0
langchain-community>=0.3.9
langgraph>=0.2.56
langsmith>=0.1.147
pinecone>=6.0.0
langchain-pinecone>=0.2.13
fastembed>=0.3.0
httpx>=0.28.0
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.0.0
streamlit>=1.40.0
